Bluecross: Use prebuilt vendor
Because vendor is now a super partition, it can't exist on both slots at the same time.
So you can't have the stock vendor just sit on both slots for the entire month because
when you flash the OTA it switches slots and the slot it switches to potentially won't
have a matching vendor image.

Change-Id: I0b4b3d8a43ecd8dbd09f3785f48559e968471a32